+++ mycodo/inputs/ads1x15.py
@@ -47,11 +47,12 @@ INPUT_INFORMATION = {
     'options_enabled': [
         'measurements_select',
         'channels_convert',
+        'i2c_location',
         'adc_gain',
         'period',
         'pre_output'
     ],
-    'options_disabled': ['interface', 'i2c_location'],
+    'options_disabled': ['interface'],
 
     'dependencies_module': [
         ('pip-pypi', 'Adafruit_GPIO', 'Adafruit_GPIO'),