+++ mycodo/inputs/sht3x.py
@@ -73,10 +73,7 @@ INPUT_INFORMATION = {
     ],
 
     'interfaces': ['I2C'],
-    'i2c_location': [
-        '0x44',
-        '0x45'
-    ],
+    'i2c_location': ['0x44', '0x45'],
     'i2c_address_editable': False,
 
     'custom_options': [